♻️ refactor: 메뉴 상세 관련 도메인 Product 영역으로 통합
