+++ src/domain/product/components/ProductList.tsx
@@ -1,10 +1,10 @@
-import { ProductItem } from './ProductItem';
-import { useRequestProducts } from '@shared/hooks/services/queries/useRequestProducts';
 import { PageContainer } from '@layout/styles/pageContainerStyle';
+import { ProductItem } from './ProductItem';
+import { useQueryProductList } from '../hooks/services/useQueryProductList';
 import { MenuList } from '../styles';
 
 export const ProductList = () => {
-  const products = useRequestProducts();
+  const products = useQueryProductList();
   return (
     <PageContainer>
       <h2>상품 목록</h2>
