+++ src/domain/product/hooks/services/useQueryProductDetail.ts
@@ -1,8 +1,8 @@
 import { useQuery } from 'react-query';
 import { fetchApi } from '@shared/utils/fetchApi';
-import type { ProductDetailData } from '@productDetail/types';
+import type { ProductDetailData } from '@product/types';
 
-export const useRequestProductDetail = (productId: number): ProductDetailData => {
+export const useQueryProductDetail = (productId: number): ProductDetailData => {
   const { data, error } = useQuery(['detail', productId], () =>
     fetchApi<ProductDetailData>({ method: 'GET', url: `/products/${productId}` })
   );
