+++ src/domain/product/pages/index.ts
@@ -1,0 +1,2 @@
+export * from './ProductDetailPage';
+export * from './ProductPage';
