--- src/productDetail/types.ts
@@ -1,17 +1,0 @@
-import { ProductData } from '@products/types';
-
-type ProductOption = {
-  name: string;
-  price: number;
-  stock: number;
-};
-
-export interface ProductOptionData extends ProductOption {
-  id: number;
-  created_at: string;
-  updated_at: string;
-}
-
-export interface ProductDetailData extends ProductData {
-  productOptions: ProductOptionData[];
-}
